+++ src/AOLPhrase.ts
@@ -1,0 +1,4 @@
+export interface AOLPhrase {
+    key: string;
+    mrk: string;
+}