+++ src/autocomplete.ts
@@ -2,6 +2,7 @@ import * as vscode from 'vscode';
 import { StrItem } from './StrItem';
 import * as https from 'https';
 import { DuckDuckGoPhrase } from './DuckDuckGoPhrase';
+import { AOLPhrase } from "./AOLPhrase";
 
 
 export function duckduckgo_ac(query: string, output: vscode.QuickPick<StrItem>): void {
@@ -9,7 +10,7 @@ export function duckduckgo_ac(query: string, output: vscode.QuickPick<StrItem>):
         method: 'GET',
         hostname: 'duckduckgo.com',
         port: 443,
-        path: `/ac/?q=${encodeURI(query)}`
+        path: `/ac/?q=${encodeURIComponent(query)}`
     }, res => {
         let resBody = '';
         res.setEncoding('UTF-8');
@@ -48,7 +49,7 @@ export function google_ac(query: string, output: vscode.QuickPick<StrItem>): voi
         method: 'GET',
         hostname: 'www.google.com',
         port: 443,
-        path: `/complete/search?q=${encodeURI(query)}&client=psy-ab`
+        path: `/complete/search?q=${encodeURIComponent(query)}&client=psy-ab`
     }, res => {
         let resBody = '';
         res.setEncoding('UTF-8');
@@ -63,6 +64,35 @@ export function google_ac(query: string, output: vscode.QuickPick<StrItem>): voi
                         /<([/]?)b>/g, ''))
                 )
             );
+
+            if (outarr.length < 1) {
+                outarr.push(new StrItem(query, 'No suggestions available'))
+            }
+            else if (query !== outarr[0].text) {
+                outarr.splice(0, 0, new StrItem(query));
+            }
+            output.items = outarr;
+        });
+    });
+
+    req.end();
+}
+
+export function aol_ac(query: string, output: vscode.QuickPick<StrItem>): void {
+    let req = https.request({
+        method: 'GET',
+        hostname: 'search.aol.com',
+        port: 443,
+        path: `/sugg/gossip/gossip-us-ura/?output=json&command=${encodeURIComponent(query)}&appid=search.aol.com&nresults=10`
+    }, res => {
+        let resBody = '';
+        res.setEncoding('UTF-8');
+        res.on('data', chunk => resBody += chunk);
+        res.on('end', () => {
+            console.log(resBody);
+            const inarr: Array<AOLPhrase> = JSON.parse(resBody).gossip.results;
+            let outarr: Array<StrItem> = new Array<StrItem>();
+            outarr = inarr.map(i => new StrItem(i.key));
 
             if (outarr.length < 1) {
                 outarr.push(new StrItem(query, 'No suggestions available'))