+++ src/search.ts
@@ -4,7 +4,7 @@ import * as config from './config';
 import { StrItem } from './StrItem';
 
 export function execute_search(input: string, extconfig: config.Config): void {
-    open_url(`${extconfig.searchEngine.URI}${encodeURI(input)}`);
+    open_url(`${extconfig.searchEngine.URI}${encodeURIComponent(input)}`);
 }
 
 export function get_suggestions(input: string, output: vscode.QuickPick<StrItem>, extconfig: config.Config): void {