+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to the "vscode-websearch" extension will be documented in this file.
 
+## 1.3.0
+
+- Added:
+  - AOL search and AutoComplete support.
+- Fixed:
+  - Properly encoding special characters.
+
 ## 1.2.0
 
 - Added: